requirements.txt
celery[redis]==5.2.7
djangorestframework
gitdb==4.0.2
GitPython==3.1.0
psycopg2
pytz
redis
requests